added Sonar

+++ source/led-blink.cpp
@@ -132,7 +132,7 @@ int main( int argc, char* argv[] )
                 //auto pulse_duration = pulseEnd - pulseStart;   // Get pulse duration to a variable
 
 
-                double distance = pulseDuration * 17150;          // Multiply pulse duration by 17150 to get distance
+                double distance = (double)pulseDuration * 17150;          // Multiply pulse duration by 17150 to get distance
                 distance = roundf( distance * 100 ) / 100;        // Round to two decimal points
 
                 /*
